requirements.txt
yfinance==0.2.32
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
plotly==5.18.0
ta==0.11.0
python-dotenv==1.0.0
